@@ -1,8 +1,6 @@
 package com.mycompany.invoise.core.entity.customer;
 
-import com.fasterxml.jackson.annotation.JsonBackReference;
-
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 public class Address {
