@@ -5,8 +5,12 @@ import com.mycompany.invoise.core.repository.InvoiceRepositoryInterface;
 import com.mycompany.invoise.core.service.InvoiceServiceInterface;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.stereotype.Service;
 
-//@Service
+@Service
+@Deprecated
+public class InvoiceServicePrefix{
+/*
 public class InvoiceServicePrefix implements InvoiceServiceInterface {
 
     @Value("${invoice.lastNumber}")
@@ -25,8 +29,8 @@ public class InvoiceServicePrefix implements InvoiceServiceInterface {
         this.invoiceRepository = invoiceRepository;
     }
 
-    public Invoice createInvoice(Invoice invoice){
-        invoice.setNumber(prefix+(++lastNumber));
+    public Invoice createInvoice(Invoice invoice) {
+        invoice.setNumber(prefix + (++lastNumber));
         invoiceRepository.save(invoice);
         return invoice;
     }
@@ -56,4 +60,6 @@ public class InvoiceServicePrefix implements InvoiceServiceInterface {
     public void setPrefix(String prefix) {
         this.prefix = prefix;
     }
+
+ */
 }
